requirements.txt
iOSbackup
gitpython
termcolor